+++ moPepGen/svgraph/VariantPeptideTable.py
@@ -1,6 +1,6 @@
 """ Varaint Peptide Table """
 from __future__ import annotations
-from typing import TYPE_CHECKING, Dict, Set, IO, List, Tuple
+from typing import TYPE_CHECKING
 from Bio import SeqUtils
 from Bio.SeqIO import FastaIO
 from moPepGen import VARIANT_PEPTIDE_SOURCE_DELIMITER, aa
@@ -9,6 +9,7 @@ from moPepGen.svgraph.VariantPeptideDict import AnnotatedPeptideLabel, PeptideSe
 
 
 if TYPE_CHECKING:
+    from typing import Dict, Set, IO, List, Tuple
     from pathlib import Path
     from Bio.Seq import Seq
     from moPepGen.params import CleavageParams
@@ -25,9 +26,11 @@ class VariantPeptideTable:
         self.index = index or {}
         self.header_delimeter = VARIANT_PEPTIDE_SOURCE_DELIMITER
 
-    def write_header(self):
+    def write_header(self, handle:IO=None):
         """ Write header """
-        self.handle.write('#' + '\t'.join(VARIANT_PEPTIDE_TABLE_HEADERS) + '\n')
+        if handle is None:
+            handle = self.handle
+        handle.write('#' + '\t'.join(VARIANT_PEPTIDE_TABLE_HEADERS) + '\n')
 
     def is_valid(self, seq:Seq, canonical_peptides:Set[str],
             cleavage_params:CleavageParams=None):
@@ -151,3 +154,17 @@ class VariantPeptideTable:
                 indices[-1] = (indices[-1][0], cur_end)
             else:
                 indices.append((cur_start, cur_end))
+
+    def sort_table(self, path:Path):
+        """ Sort peptide table """
+        with open(path, 'wt') as handle:
+            self.write_header(handle)
+            for seq in self.index:
+                annos = self.load_peptide_annotation(seq)
+                for anno in annos.values():
+                    for seg in anno.segments:
+                        subseq = str(seq[seg.query.start:seg.query.end])
+                        line = '\t'.join([
+                            str(seq), anno.label, subseq, seg.to_line()
+                        ]) + '\n'
+                        handle.write(line)