+++ CHANGELOG.md
@@ -10,6 +10,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [1.5.0-rc2] - 2025-06-21
+
+- Updated `mergeFasta` to also merge the peptide table if exists.
+
+- Updated `callVariant` so the peptide table is sorted and the order of peptides will sync with the fasta file.
+
 ## [1.5.0-rc1] - 2025-06-17
 
 - Added `--codon-table` and `--chr-codon-table`. The former sets the codon table to use, and the latter overrides it for specific chromosomes.